+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 Unreleased
 
 - Added the ability to plot several types of graphs, including boxplot and area plots.
+- Added `--grid` and `--no-legend` options.
+- Added `--marker` and `--linestyle` options in the line graph.
+- Added `--alpha` and `--bins` options in the histogram.
+- Added `--stacked` option in the bar plot.
 - Added support for changing the style of the graph.
 
 ## 0.1.0